Released version 0.11.1 via File Upload
Version 0.11.1 - All NWS Current Observation XML files *should* now open without issue. The "Open Forecast XML File" button is merely a placeholder at this time, as the viewing of NWS Forecast XML files has not been implemented yet. Planned features and ideas for future versions are now noted in the new "TO DO" section of the file. The last release on GitHub was version 0.9.1; 0.10.x was never released. I still plan on continuing to work on this tool here and there. Please note that I may consider rewriting this tool in the future in a different language.